[fix] : conflict 해결
