--- src/main/java/com/dia/dia_be/service/pb/impl/PbPbReserveServiceImpl.java
@@ -1,28 +1,0 @@
-package com.dia.dia_be.service.pb.impl;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import com.dia.dia_be.dto.pb.ReservesDTO.ResponseReserveDTO;
-import com.dia.dia_be.repository.ConsultingRepository;
-import com.dia.dia_be.service.pb.intf.PbReserveService;
-
-@Service
-public class PbPbReserveServiceImpl implements PbReserveService {
-
-	private final ConsultingRepository consultingRepository;
-
-	public PbPbReserveServiceImpl(ConsultingRepository consultingRepository) {
-		this.consultingRepository = consultingRepository;
-	}
-
-	@Override
-	public List<ResponseReserveDTO> getApprovedReserves(boolean status) {
-		return consultingRepository.findConsultingsByApprove(status)
-			.stream()
-			.map(ResponseReserveDTO::from)
-			.toList();
-
-	}
-}
